chore: Add custom pagination and users list
A custom pagination was added to support page and limit as required,
default pagination supported by drf do not have them named page and
limit. This nedded a custom paginator to satisfy the requirements.

The fetching of all users was implemented by adding
mixins.ListModelMixin and it was added to the now renamed Viewset called
UserCreateAndListViewSet and this is becasue logically speaking,
retrieving a list should not be added to a detail view.